+++ app/models/ability.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 class Ability
-  include Valhalla::Ability
+  include Hydra::Ability
   # Define any customized permissions here.
   def custom_permissions
     alias_action :show, :manifest, to: :read
@@ -204,6 +204,44 @@ class Ability
 
   def universal_reader?
     current_user.curator? || current_user.image_editor? || current_user.completer? || current_user.fulfiller? || current_user.editor? || current_user.admin?
+  end
+
+  def read_permissions
+    super
+    can :read, Valkyrie::Resource do |obj|
+      valkyrie_test_read(obj) || valkyrie_test_edit(obj)
+    end
+  end
+
+  def edit_permissions
+    super
+    can [:edit, :update, :destroy], Valkyrie::Resource do |obj|
+      valkyrie_test_edit(obj)
+    end
+  end
+
+  def valkyrie_test_read(obj)
+    group_readable?(obj) || user_readable?(obj)
+  end
+
+  def group_readable?(obj)
+    (user_groups & obj.read_groups).any?
+  end
+
+  def user_readable?(obj)
+    obj.read_users.include?(current_user.user_key)
+  end
+
+  def valkyrie_test_edit(obj)
+    group_editable?(obj) || user_editable?(obj)
+  end
+
+  def group_editable?(obj)
+    (user_groups & obj.edit_groups).any?
+  end
+
+  def user_editable?(obj)
+    obj.edit_users.include?(current_user.user_key)
   end
 
   class NilToken
